serial: stm_asc: Cache interrupt enabled register
This is a micro-optimisation copied from the 8250 driver. In many
cases modifications of the interrupt enables register (INTEN)
involve a read-modify-write, and of these many of these actually
don't change the value being written.

So cache the value of the register in the ASC's data structure, thus
avoiding many accesses to the hardware register.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>