mmc: hack to use the Micron IZAI8 JW704 8GB on Lille
This actually is a dirty hack to use the Micron IZAI8 JW704 8GB)
on Lille board. This card fails when perform the bus test step
and it is not able to switch to 1bit mode.
This card has been used for testing the boot from eMMC
on B2039 so it's worth having it working on our platforms.
I suspect this is a card issue!

The extra delay is not a problem for other cards fortunately.
It has been also verified that no other cards (Toshiba, SanDisk
and old Micron) suffer of this problem.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>