stm: i2c: Allow fastmode to be initialised from platform data
Add a flag to the stm I2C platform data which can be used to initialise
the device's fastmode flag. This allows boards which have an I2C bus
populated entirely with fastmode capable slave devices to set the
flag as part of their initialisation of the I2C device, and so remove
the need for later configuration using the sysfs device.

This models the way the PXA I2C driver sets its fast_mode flag.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>