stm_spi_fsm: fix race condition in fsm_wait_busy loop
This patch fixes what appears to be a race condition in the fsm_wait_busy()
loop.  Previously, we attempted to read the FIFO whenever it was reported data
was available.  This allows the CPU to read from the FIFO in parallel with the
controller writing any remaining data.  In most cases this seems to work well.
However, while testing at low clock rates (~2MHz), it was found that the
fsm_wait_busy() loop would fail to complete.  If instead, we wait for the
sequence to complete before reading the FIFO, then the problem would go away.
Although the issue is not properly understood, and only appears at stupidly slow
clock rates, the workaround has little impact on performance in the case of
fsm_wait_busy(), so it seems sensible for it to be included.

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>