stm: b2057: use stm-nand-bch driver for NAND support
In most cases, it makes sense to use the FLEX NAND driver by default, since it
works on all systems, and does not interfere with the manufacturer-programmed
bad block markers.  However, the b2057 board is populated with a NAND device
that is only supported by the stm-nand-bch driver (due to multi-bit ECC and
Page/OOB geometry).  This patch updates the board setup and defconfig to select
the stm-nand-bch driver.

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>