stm-nand-bch: add private page cache for non-aligned read operations
This patch adds support for a private page cache when performing non-aligned
data reads, in a similar way to the other stm-nand drivers.  The main motivation
is to make up for deficiencies in the mtdblock layer which employs a fixed
sector-size of 512 bytes, regardless of the properties of the underlying Flash
device.  In the case of NAND Flash, this can lead to each page being re-read
unnecessarily many times.

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>