pcie_stm: Use slot 1 instead of 0 for first device
One of the changes introduced in commit 41eb92a7384a2bd60214 was that
the first device on the PCIe bus was labelled as device zero rather than
one. This turned out to be a really bad idea, as some devices object
strongly to be labelled as device zero since this is usually reserved
for the root complex itself. This patch reverts to labelling the first
device as 1 rather than 0.

Signed-off-by: David McKay <david.mckay@st.com>