sh4_stm: extend cpu_info with family variant information
Extend the cpu_info with family variant information for the
SH4 family, specifying if we are running on an SH4-202 or ST40-300
and export this information to the user space through the aux vect
in the AT_BASE_PLATFORM entry.
The actual variant is detected at runtime by probing the CCN_PVR
register (as done for cpu family and cache system).

C runtime ld.so will use it to load implementation specific
libraries for optimisation.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Reviewed-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
[stuart.menefy@st.com: Added SH4-101/3 and renamed SH4-300 to ST40-300]
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>