pm: enhance the wake-up support adding WoL+
This patch adds a new check for the devices that can wake-up the system.
In fact, now we have the IC+ 101A/G PHYs that can wake-up the
whole system from a magic packet.
In this case, the SoC suspend core has to keep enable only the PHY
clock because the MAC can be totally powered-off.
The new field is called: stm_phy_can_wakeup

V2:	As Stuart suggested: renamed the misleading
	name previously adoped: only_phy_can_wakeup
	as stm_phy_can_wakeup.
	Unified the names eth_phy_can_wakeup and eth1_phy_can_wakeup:
	stm_mac0/1_can_wakeup.
	Added a new __check_mdio_wakeup_device for MDIO.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Acked-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Cc: Stuart Menefy <stuart.menefy@st.com>