stm: ilc3: Workaround interrupt arriving during suspend
This patch adds the suspend callback to the ilc in order to
enable the wakeupable irq lines before the cpu is suspended.
This works around a problem seen when suspending a system and
an interrupt arrives during the suspend process.

This is required because the ilc driver does not provide a
ilc_irq_disable, therefore the kernel virtually disables the irq lines
(marking them as IRQ_DISABLED), but keeps them enabled at the hardware
level.

Looking at the suspend code:

1. dpm_suspend_start(PMSG_SUSPEND);
2. dpm_suspend_noirq(PMSG_SUSPEND);
3. arch_suspend_disable_irqs();
4. sysdev_suspend(PMSG_SUSPEND);
5. suspend_ops->enter(state);

In (1) all the devices are suspended and the wakeupable devices
prepare the wakeup-irq.
In (2) the kernel disables all the irq lines but as described above,
for ilc interrupts the irq lines are virtually disabled (but physically
still enabled).

Under stressing condition (i.e. stmmac irq: ping -i.2 x.x.x.x) the
wakeup-irq could be raised after (1) but before (3) in which case the
kernel masks the irq in handle_level_irq() masking the interrupt in
hardware, and loosing the wakeup capability the irq-line had.

The new ILC suspend callback, which is called from (4), enables again
all the wakeup-able irq lines to guarantee the system can wake-up.

Reference: Bugzilla 17159

Signed-off-by: Nunzio Raciti <nunzio.raciti@st.com>
Acked-by: Francesco Virlinzi <francesco.virlinzi@st.com>