hwmon: stm-temp: added correction_factor platform_data
This patch adds a new field 'correction_factor' in the platform_data
to allow a chip specific correction to the value read on hw.
This is required because the standard correction (20) isn't valid for
newer chips, for example the STxH205.
Now each chip will have to declare the correction it needs.

Reference: [Bugzilla 21489]

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>