stm_nand_bch: update Page-OOB mapping
The stm-nand-bch driver implements a mapping process whereby the "interleaved"
data and ECC sectors are transposed to the more traditional layout of data in
the Page area and ECC in the OOB area.  This allows us to maintain some level of
consistency for the "mtd_oob" operations and to support some of the user-space
tools, such as "nanddump".

Previously, the mapping process would divide the OOB area according to the
number of data sectors within the page.  Typically, the entire OOB would be
consumed by the ECC data for each sector.  However, on devices with larger OOB
areas, each OOB sector would include some "free" space after the sector's ECC
data.  Based on the assumption that the "free" space is not made available for
use, the current implementation would simply discard any data in the "free"
area.

As it turns out this approach can lead to some subtle problems when using the
driver in unusual ways.  One such example occurs when we are using the
mtd_read_oob() support to perform "raw" reads of "BCH30" data when the driver is
configured for "BCH18"; the "BCH18" mapping process ends up discarding the
additional "BCH30" ECC data.

In most cases, this situation does not arise.  However, one of the fundamental
requirements of the STM NAND Tools package is that it must be able to read and
write data in any of the supported ST ECC schemes regardless of which driver or
configuration is currently in operation.

Rather than extend the existing "OOB divided into equal sectors" approach, we
instead adopt a simpler mapping process for ensuring all the OOB data gets
returned.  The ECC data for each sector is mapped to a contiguous area at the
start of the OOB, with any "free" space mapped to the remainder of the OOB.

This change has no impact for the normal use of NAND, only for the way in which
"mapped" data is interpreted.  In most cases, the interpretation is done by an
engineer during debugging; for example, viewing data returned by "nanddump".  As
far as I am aware, the only software component that attempts to use the mapped
data is the STM NAND Tools package, and this packages has already been updated
to support the new mapping scheme since "stm-nand-tools-0.6.0" (still
pre-release).

Signed-off-by: Angus Clark <angus.clark@st.com>