sh: pmb: Fix checks for existing uncached mapping
The code which checks for an existing uncached mapping has a number
of issues. Firstly the code which reads the PMB registers discards the
valid bit, which is not only unsafe but also means the comparison can
never succeed. Secondly it only works in the case where exactly one
PMB entries is required by the uncached mapping which was true in the
past but may not be if CONFIG_PMB_LARGE_UNCACHED_MAPPING is set.
Similarly the req_pos parameter to pmb_calc can cause a buffer
overrun when multiple PMB entries are required.

Fix __get_pmb_entry, generalise the check for an existing uncached
mapping when there are multiple PMB entries, and change the req_pos
parameter to pmb_calc to be an int, as the only code which needs to
know which PMB entries were actually allocated now gets that by parsing
the existing data structures.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>