stm: hom: sh: Make board support aware of wakeup_device
This patch adds the required code to pass the HoM board support code
information about which devices can wakeup the system.
This is done because based on which devices can wakeup the system
some board power lines may or may not be turned off.

Update all the boards which use this API so that they hold the
Ethernet PHY in reset when not using wake on LAN.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>