Register RTC handler for e2 procfs
This registers two handlers for the RTC provided by aotom. The e2 gui uses
the /proc/stb/fp/rtc file for reading and writing the RTC. In the past there
was no connection between the RTC part in aotom and the e2 procfs.

The RTC needs localtime instead of UTC otherwise the display shows the UTC
time which is maybe several hours off.

Signed-off-by: Christian Ege <k4230r6@gmail.com>